fix(migration): change timestamp fields to bigInteger in money_transactions table

+++ database/migrations/2026_05_22_173613_money_transaction.php
@@ -66,9 +66,9 @@ return new class extends Migration
             |--------------------------------------------------------------------------
             */
 
-            $table->integer('createdAt')->nullable();
-            $table->integer('updatedAt')->nullable();
-            $table->integer('deletedAt')->nullable();
+            $table->bigInteger('createdAt')->nullable();
+            $table->bigInteger('updatedAt')->nullable();
+            $table->bigInteger('deletedAt')->nullable();
 
             $table->foreignUuidFor(User::class, 'createdBy')->nullable();
